refactor: replace new onboarding
