+++ apps/dashboard/app/new/components/onboarding-success-step.tsx
