+++ apps/dashboard/app/new/hooks/use-workspace-step.tsx
@@ -6,7 +6,7 @@ import { zodResolver } from "@hookform/resolvers/zod";
 import { StackPerspective2 } from "@unkey/icons";
 import { Button, FormInput } from "@unkey/ui";
 import { useRouter, useSearchParams } from "next/navigation";
-import { useRef, useTransition } from "react";
+import { useRef, useState, useTransition } from "react";
 import { useForm } from "react-hook-form";
 import { z } from "zod";
 import type { OnboardingStep } from "../components/onboarding-wizard";
@@ -31,6 +31,7 @@ type WorkspaceFormData = z.infer<typeof workspaceSchema>;
 
 export const useWorkspaceStep = (): OnboardingStep => {
   // const [isSlugGenerated, setIsSlugGenerated] = useState(false);
+  const [workspaceCreated, setWorkspaceCreated] = useState(false);
   const formRef = useRef<HTMLFormElement>(null);
   const router = useRouter();
   const searchParams = useSearchParams();
@@ -79,6 +80,7 @@ export const useWorkspaceStep = (): OnboardingStep => {
   const createWorkspace = trpc.workspace.create.useMutation({
     onSuccess: async ({ workspace, organizationId }) => {
       workspaceIdRef.current = workspace.id;
+      setWorkspaceCreated(true);
       switchOrgMutation.mutate(organizationId);
     },
     onError: (error) => {
@@ -110,6 +112,10 @@ export const useWorkspaceStep = (): OnboardingStep => {
   });
 
   const onSubmit = async (data: WorkspaceFormData) => {
+    if (workspaceCreated) {
+      // Workspace already created, just proceed
+      return;
+    }
     createWorkspace.mutateAsync({ name: data.workspaceName });
   };
 
@@ -172,7 +178,7 @@ export const useWorkspaceStep = (): OnboardingStep => {
               // }}
               required
               error={form.formState.errors.workspaceName?.message}
-              disabled={isLoading}
+              disabled={isLoading || workspaceCreated}
             />
             {/* <FormInput */}
             {/*   {...form.register("workspaceUrl")} */}
@@ -189,14 +195,19 @@ export const useWorkspaceStep = (): OnboardingStep => {
     kind: "required" as const,
     validFieldCount,
     requiredFieldCount: 1,
-    buttonText: "Continue",
-    description: "Set up your workspace to get started",
-    onStepNext: () => {
-      if (isLoading) {
-        return;
-      }
-      formRef.current?.requestSubmit();
-    },
+    buttonText: workspaceCreated ? "Continue" : "Create workspace",
+    description: workspaceCreated
+      ? "Workspace created successfully, continue to next step"
+      : "Set up your workspace to get started",
+    onStepNext: workspaceCreated
+      ? undefined
+      : () => {
+          if (isLoading) {
+            return;
+          }
+
+          formRef.current?.requestSubmit();
+        },
     onStepBack: () => {
       console.info("Going back from workspace step");
     },
